Nonrecursive version of aFiles::rmRf needed more love to work properly. It was failing to remove things, but it wasn't removing any inappropriate things, so this is not the end of the world
git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4548 c89fab08-2f9b-42bf-8176-bd3ad6bee547

+++ lib/toolkit/BaseaFiles.class.php
@@ -358,8 +358,7 @@ class BaseaFiles
   {
     $originalPath = $path;
     $paths = array();
-    error_log("Removing $path");
-    $paths[] = $path;
+    $paths[$path] = $path;
     while (count($paths))
     {
       $path = array_shift($paths);
@@ -371,18 +370,31 @@ class BaseaFiles
       if (aFiles::statIsDir($stat))
       {
         $list = aFiles::ls($path);
-        foreach ($list as $file)
+        // If the directory has contents, add them to the list, and make sure they are
+        // removed before the parent folder is. This can happen more than once if they
+        // have nested contents, which is OK
+        if (count($list))
         {
-          $filePath = "$path/$file";
-          if (strlen($filePath) < strlen($originalPath))
+          foreach ($list as $file)
           {
-            throw new sfException("I almost tried to delete something higher up than the original, I don't like this, bailing out");
+            $filePath = "$path/$file";
+            if (strlen($filePath) < strlen($originalPath))
+            {
+              throw new sfException("I almost tried to delete something higher up than the original, I don't like this, bailing out");
+            }
+            $paths[$filePath] = $filePath;
           }
-          $paths[] = $filePath;
+          // Make the parent the last path in the queue to be removed, removing any duplicate from the list first
+          unset($paths[$path]);
+          $paths[$path] = $path;
         }
-        if (!aFiles::rmdir($path))
+        else
         {
-          return false;
+          if (!aFiles::rmdir($path))
+          {
+            return false;
+          }
+          $removed[$path] = true;
         }
       }
       else
